Use fully qualified lakeFS URIs as inputs to file operations
Instead of saving state within the filesystem, we change the approach
to operate on lakeFS file URIs of the form lakefs://<repo>/<ref>/<resource>.

To do this, a `parse()` method was added to parse such input paths via
regex matching, and fail if either of the three components above is
missing or malformed. Test coverage was added for the regexes on repo
and branch names, respectively.

The parsing step takes the place of raising exceptions for missing refs
coming previously from the `ref=None` keyword arguments to almost all
file system APIs. These keyword arguments, along with the `repository`
member of the file system, are now gone, to achieve maximum flexibility
with respect to repos and branches using a single client.