Implement lakeFS-specific info() method
Since lakeFS never returns paths terminated with slash, and all paths
are file resources ("objects") specifically, we can replace info() with
a simpler implementation.

In the current state, superfluous regex parsing takes place in the info(),
ls() and exists() methods, which are almost exclusively used internally.
A concept for passing the unpacked parsed inputs will be explored in a
follow-up.