Add random file factory for easily creating files
In tests where files need to be uploaded or hash-checked, we need a way
to create files with random names and contents, so adding a small class
for it like the API counter for instrumenting seems like a good thing to
do.

Currently the random file factory is function-scoped like the temp path,
and not session-scoped like the temp path factory fixture.