Place commithook in top-level `put` and `rm` methods
Doing so avoids commit spam in recursive uploads or deletes. It is
assumed that the requested resource is always sufficient for creating a
meaningful commit message.

The `rm` and `put` implementations added are immediately deferring to the
super calls, and then optionally firing the commit hook as before.

The default hook was changed to allow both `rm_file` and `rm` events.