Improve exceptions, add translator context manager
Since the API exceptions should not be raised to the user, who expects
generic IO exceptions from Python builtins, we can translate them with
e.g. a context manager. A prototype was added here.

For now, some of the user exceptions were refactored and concretized to
better reflect the assumptions to the lakeFS backend.