Split off release job into a new file
Since the scope of running a release job is different enough, we put it
into a new action file.

Triggers are now either manual or a published release. The rest is just
the same as before.

Rename the testing job to "test" to better reflect its purpose and
differentiate it to the build job.