Change lakeFS file system instantiation mechanism to take raw arguments
Instead of an instantiated client, take the raw lakeFS configuration
arguments to construct the lakeFS file system. This has the distinct
advantage that no `lakefs_client` imports need to happen on the user's
part.

For e.g. `pandas` usage, this has the effect that all attributes (e.g.
host, username) can be given directly into the `storage_options` dict.
The readme was updated to showcase this.

Also, instead of a direct config, environment variables were defined,
which can be set to use zero-code config for the file system, and thus
skip `storage_options` altogether.

All tests were updated to use the new mode of instantiation. In the
future, we can switch to envvar-config and drop the fixture configuration
entirely (except for a unit test verifying correct behavior).