Assert instance caching works by adding a test for it
When we passed in a lakeFS client, it was up in the air whether it would
support instance caching properly. But now that we switched to the raw
config, it is straightforward to check.

Adds a test checking the cache on the key (`_fs_token` attribute) and
the cache size when initializing a new instance.