Improve flexibility of commithooks by adding more arguments (#36)
* Improve flexibility of commithooks by adding more arguments in a context

Adds the repository, the branch, and the diff that is being committed as
context arguments to the commit hook.

A context is a simple named tuple made directly from the arguments used
in the file system event calls.

The file system event object was changed from a raw string to an enum.

Also updates the docstring for the `CommitHook` type. Arguments were
renamed from `event` to `fs_event`, and `rpath` to `resource`,
respectively, to avoid confusion. The butchered rendering in Pycharm
was fixed by line-breaking on proper paragraphs, although it does
disrupt the line-length config (?).

* Change fs event enum type to `StrEnum`, fix print statement

Now, the actual lower-case file system event name shows up again in the
error message.

* Use explicit quotes instead of repr shorthand