Add test for file system initialization methods
Since we now have three rivaling modes of initialization for the file
system, it is time to test the priority order is correct.

This brought up a quirk related to instance caching: Since both the
`lakectl` config file and environment variables are read implicitly,
they do not require arguments, and thus the first method used between
config file and envvar initialization will cause cache hits, and
suffocate the other method unless the cache is cleared.