Clear instance cache everytime in file system fixture, update README
We fell victim to implicit caching in the tests, where a) an API-counted
test was not working due to the patch not being picked up (or, rather,
a client being patched twice), and b) a filesystem used `postcommit`
only through a cached instance, and not in a scope.

Since we set the file system as a test-scoped fixture, we should clear
the cache every time before initializing.

Also, instead of the private dict cache clear, prefer the `clear_instance_cache`
public method.