Call `objects_api.stat_object` endpoint in case of file in `fs.info`
It is wasteful to always get the object list from `objects_api.list_object`,
and might be slower for buckets/repos with lots of objects.

Hence, if we decide that the input to `fs.info` is a file (heuristic:
input does NOT end on slash), we call the `objects_api.stat_object`
endpoint instead.

Judging from an earlier debugging session, this also saves some more
API calls that do not happen as ripple effect from `fs.ls`.