Add `lakefs_spec.client_helpers` module for hook snippets
Since we want the user to not have to write too much code, we provide a
sort of recipes / snippets module that gives some of the most-used
client operations as Python functions to call.

In the particular case of post-operation commits, which were a one-liner
previously, the usability has gone down (or become more difficult, rather),
but the extensibility of post-file system-event operations is greatly
increased.