Add test for `fs.cp_file`
Since the copy should be made to a new branch, create a context manager
fixture for temporary branch creation.

This is superfluous code with the `temp_branch` fixture, but I have no
better idea to unify the two right now.

The context manager is also untyped for now, since I was unable to get
the typing correct in reasonable time.