Add directory listings caching to `LakeFSFileSystem.ls`
Caches results of `fs.ls` calls in the class' `DirCache` instance.

As an added bonus, if a file is requested by `fs.ls` that has a cache
entry for its parent, but is not yet found in the cache, this implementation
will fetch those files and append them in the correct place in the
existing cache entry. This "hot-reloading" of cache entries is a little
more robust than just raising an exception and leaving the user to
manually list the objects again.