Opt to extend the cache key instead of updating individual objects
The reason is that with the present caching logic, code that reaches this
section and encounters a cache key for the parent dir cannot contain any
information for the requested resource (prefix in `objects_api.list_objects`),
and thus any info obtained through the `list_objects` call cannot be
present in the cache key.

This means it is safe to extend the entry regardless.

Add a test to check the `detail=False` case both cached and uncached.