Fix test failure by parsing prefix from `list_objects` result
As discussed on the GitHub PR #86. Since we have no way of knowing
whether a given prefix is a file or a directory, we are instead taking
the answer the backend gives us.

Nested folders are NOT written to their respective cache entries, since
that would create hits for resources that have not been requested yet.