Change `get_file` implementation to fetch byterange via API call
This was changed in the lakeFS SDK from the buffered reader used before.

Now, we need to fetch byte ranges, which we were doing in the
`LakeFSFile` already, with every `objects_api.get_object` call.

The main difference, however, is that we cannot just pass increments of
the file system's chunk size, because that will create HTTP416 errors
(request range not satisfiable).

Instead, first fetch the object size from the backend via `fs.info()`,
then guard against the overflow by max'ing it with the next increment
of the chunksize.

We also change a test to purposefully require multiple `get_object`
range requests to see if the logic works as intended.