Change blockstore tests to mock lakeFS API return objects
Since the `get_blockstore_type` method was inlined (and removed), we
now have to change the mock methods.

We do this by defining mock namedtuples that replicate the API models
that we are testing for. Despite not having the pydantic machinery
available, it is sufficient for our current test setup.