deps!: Remove source annotations from lockfiles
This is needed since platform-specific dependencies
affect these annotations, depending on if a package
has been explicitly depended on or is a platform-specific
transitive dependency (e.g., `appnope`, which is required
directly by us, but also conditionally by `ipykernel`).
This would break CI due to the platform differences between
dev and runner platforms.