Explicitly check for existence in `LakectlConfig.read()`
Based on the class responsibility model, the `LakectlConfig` class
should not assume that existence of the input file has already been
checked.

However, in the file system case, this is necessary to gracefully decay
into an empty config in case the default config does not exist, or cannot
be read.