Implement lakeFS transactions with versioning helpers
Introduces a subclass of fsspec's builtin transactions called
`LakeFSTransaction`, which in addition to its normal use has a number of
predefined common versioning operations.

As the transaction is supposed to be used within the context manager
block, the generator signature was changed to return the transaction
instance, which is a departure from the fsspec instance.

When versioning operations are requested, they are forwarded to their
respective client helper function along with their supplied arguments,
and the stack of file objects and versioning ops is unwound in the
context manager's exit method.