Switch `put_file` implementation to defer to super, add failure test
I thought I checked in the `put_file` change to the base class
implementation, but that was apparently not the case.

Changes all `put_file` operations in the tests to pass `autocommit=True`
to defer the file uploads and add to the transaction list.

Also adds a test verifying that versioning operations are cancelled on
an exception, using the example of a commit.