Remove hook functionality
The hook concept has been all but superseded by the new transaction API,
which enables users to define a versioning workflow more naturally and
make fine-grained decisions on when to do versioning, on which files
etc.

Since the hooks are now only relevant for lower-level instrumentation,
and we currently have no users using them, we remove them again. In a
separate PR, we updated the demo notebook to use transactions everywhere,
too.