Remove references to hooks from documentation, advertise transactions
Even though we only have the predefined client helpers on the
transaction, which loses flexibility from the hook concept, the
workflow using transactions is so superiorly readable and intuitive that
most, if not all users should not be hindered.

Advertises transactions explicitly as the tool to do versioning flows
with.