Refactor transactions to submodule, introduce placeholders for commits
Similarly to `fsspec`, the transaction now lives in the
`lakefs_spec.transaction` submodule.

Introduces the `Placeholder` generic type for a deferred value that is
only available after transaction completion.

This is important for values that are not available immediately after
execution of a transaction helper, such as a commit SHA, which has to
be computed first by the backend, and this happens only on transaction
completion.

To be able to still pass these values to other consumers in the
transaction, the `ref` argument was changed to also take a `Placeholder`
as an argument.

The placeholder is tracked by the transaction's file+op queue, and
populated with the actual return value of the executed underlying
client helper after the client helper operation completes.

In subsequent operations consuming these placeholder values (with now
populated values), the placeholder values are unpacked before the client
helper execution, which keeps the client helper interface clean.

These details are hidden away from the user, who just proceeds to use
commit SHAs as normal. However, printing these placeholders during the
transaction will not show the value for obvious reasons.