Limit training data to 2010 and 2020, change model to random tree classifier
Also update all filenames and commit messages to reflect the change in
tutorial scope.

This cuts down docs execution time to under 10 seconds, so it is better
for frequent documentation rebuilds while developing.

Additionally, we inline the requirements installation into the notebook,
since that is the right place for them. The prerequisites part and
virtual environment guide is removed, since users are expected to know it.