Add `exist_ok` to all resource-creating client helpers
This ensures that usage can be made idempotent, which (for us) has the
tangible advantage of allowing for faster docs development iterations.

The heuristic to detect a backend "exists" error is only the status code
409 for now, though we could also do a substring search on the reason.
This however would incur a JSON parse.

Changes the tag creation API to return a `Ref` model object.

The `ensure_branch` was renamed to `create_branch` on account of the
creation option `exist_ok` that was added.