Change certain aspects of the data science tutorial notebook
Improves a few aspects of the tutorial:

* The weather API data is now saved to a temporary file, no more
directories are created. The tempfiles are closed immediately after
their last uses in the upload transactions.
* The separate `put` operation section was cut, because the showcase of
a commit abort is unnecessary. Instead, the focus is shifted to the
transactional upload of the data frames.
* Instead of shape comparisons, `pandas.DataFrame.equal` is used to
establish equality of local vs. remote downloaded data. This actually
caught a dtype mismatch between two datasets.
* Through style changes and regroups, a number of text passages were
slimmed down or cut entirely.