Change `fs.info()` logic to recognize non-slash-terminated directories
Previously, our heuristic for calling `objects_api.stat_object` was the
resource terminating in a slash.

However, we put this into an `if-else` branch, meaning that
`objects_api.list_objects` would never be called on a directory resource
that was not slash-terminated. This leads to erroneous "not found"
errors on existing directories.

The heuristic now changes to the strategy "call `stat_object` first,
return if successful, otherwise call `list_objects`" strategy. This
ensures that on an unsuccessful `stat_objects` call on a directory, the
correct `list_objects` call follows directly afterwards.

Also adds a regression test for the behavior.