Fix `fs.get()` and `fs.put()` by expanding path with versioning prefix
Otherwise, parsing the lakeFS URI (with or without scheme) fails since
the repository and ref were stripped.

Also, since the directories are part of the expanded path list by
default in the base implementation, which is designed to make a dir
on an `lpath` directory input in `put_file`/`get_file`, we need to check
that any input to our custom implementations does not trigger the
`precheck` checksum verification branch. This is done by checking if
the lpath input is a directory.