Allow leading lakefs:// scheme in input paths (#169)
Since the scheme is a part of the URI, it should be allowed in the
parsing regex. This is achieved by adding an optional positive lookahead
capture group.

This has the benefit that the scheme, since we don't need it, does not
appear in the regex match groups, so all other code can stay unchanged.

Adds a regression test for scheme parsing.