Remove tempfile machinery from `urlretrieve` calls
`urlretrieve` itself uses tempfiles under the hood, which explains the
Windows error as it tries to open the passed filename in write mode,
which is impossible on an already opened tempfile on Windows.

Instead, we leave out the filename, prompting `urlretrieve` to use its
own tempfile, and clear out all created tempfiles after the weather data
has been downloaded.