Change certain aspects of the data science tutorial notebook (#158)
* Change certain aspects of the data science tutorial notebook

Improves a few aspects of the tutorial:

* The weather API data is now saved to a temporary file, no more
directories are created. The tempfiles are closed immediately after
their last uses in the upload transactions.
* The separate `put` operation section was cut, because the showcase of
a commit abort is unnecessary. Instead, the focus is shifted to the
transactional upload of the data frames.
* Instead of shape comparisons, `pandas.DataFrame.equal` is used to
establish equality of local vs. remote downloaded data. This actually
caught a dtype mismatch between two datasets.
* Through style changes and regroups, a number of text passages were
slimmed down or cut entirely.

* Close tmp files after creation

to enable writing to the temporary files on windows. Set delete=False to
persist the temporary files in memory after closing. Later they are
deleted via os.remove().

* (tutorial) Trim lakeFS setup
as the setup is part of other tutorial pages.

Changed some spellings of lakeFS to a lowercase l to be consistent with
how they spell themselves.

Moved pip installation of tutorial notebook packages from magic cell to
markdown as the installation via pip from the latest PyPI release might
deviate from the current developer state (but a user will only see the
latest PyPI released docs and thus install the right package).

* Consistify lakeFS-spec spelling

* Readd package installation cells

such that the packages are installed in the CI pipeline. The
installation of the lakeFS-spec package is not in an executable cell but
only referred to in a markdown cell. This circumvents errors in develop-
ment due to an outdated revision installation via PyPI.

* Wording changes and nbstripout

* Add `--upgrade` to pip install instruction, fix small wording problems

* Remove tempfile machinery from `urlretrieve` calls

`urlretrieve` itself uses tempfiles under the hood, which explains the
Windows error as it tries to open the passed filename in write mode,
which is impossible on an already opened tempfile on Windows.

Instead, we leave out the filename, prompting `urlretrieve` to use its
own tempfile, and clear out all created tempfiles after the weather data
has been downloaded.

---------

Co-authored-by: Max Mynter <maxmynter@me.com>