Add HTTP409 check in `create_repository` for forward compatibility
With lakeFS 1.2.0. This breakage was observed in a local test against
`treeverse/lakefs:latest`, which emits HTTP409 on an existing repo
instead of HTTP400.

The code assumes that HTTP400 might happen for reasons other than an
existing repository, but HTTP409 does not.

Since there is no easy way to obtain the version from the backend, we
end up with an if-branch on the status code.