fix(ci): Record test coverage on fsspec package
In order to gauge the effectiveness of our test suite, we should not
only measure coverage on our own code, but rather on the abstract
base classes for the file system and file implementations as well.

This commit adds the necessary parameters to the `pytest` invocation
in CI, which then ends up in the CodeCov reports.