Modify error translator to raise more informative errors
Prepares the `status`, `reason`, and `filename` arguments unconditionally
for each observed API error.

This way, users who see translated errors actually know what the filename
involved is instead of just getting the (less than ideal) lakeFS API
reason.

One particular behavior is that of `fs.ls()` in case the branch/ref is
not found: Here, unlike in the repository case, the branch is not part
of the error message, which means just a generic "file not found" error.

We use OS Errnos for more exotic statuscodes, and HTTP codes for 401,
403 and 404 errors.