Collapse status, reason and filename into one string instead
Removes the extra "[Errno xyz]" prefix from `FileNotFoundError`s.

Also, this presents the HTTP status in front of the reason, which is
parsed out of the message body and not from the `reason` attribute,
because the former is sometimes more informative than the latter, for
example in case of a non-existent repository.