fix: Correctly update dircache from ls
The `ls` command was not updating the dircache correctly, which could
lead to incorrect results when listing the same directory multiple times.

In particular, the implementation would always add all results from the
API call to the dircache for the directory, regardless of whether the results
were already in the cache entry (essentially doubling the cache entry for each
subsequent call).

This behavior would only be triggered by passing `refresh=True`, which
bypasses the shortcut evaluation that returns a dircache entry directly.