fix: `ls` is no longer recursive by default, cache invalidation in `rm`
The `ls` function in its previous implementation would always return
items in subdirectories recursively, which is not an intuitive behavior.

This commit introduces an optional `recursive` kwarg, which is set to
False by default.

Additionally, this commit fixes the invalidation of the dircache in `rm`
and adds appropriate test cases.