tests: Exclude method overloads from coverage
This prevents type stubs used for method overloads from being included in the
coverage report, since the cannot be actually called and are only used for type checking.

In turn, this should reduce the number of uncovered lines in the changeset reported
by CodeCov for the code executed as part of the CI pipeline.