Fix test flakes due to accidental `main` branch resets
This was observed in local tests, but apparently not in CI. The culprit
was that in tests, we attempted to reset the `main` branch instead of the
`temp_branch`, which in some versions apparently is not allowed.

Hence, to be safe, we try out all resets on the `temp_branch` returned
by the same-name fixture instead.

Co-authored-by: Adrian Rumpold <a.rumpold@gmail.com>