Overload `LakeFSFile.__del__` to pass statement
Previously, the file was closed again on deletion, which is bad practice
if an exception occurred in write mode, since any exception on write
will immediately reappear in the exit routine.

This has potential to disturb Python exit cleanup, so we just unset the
closing attempt in __del__. Any users wanting to upload a file implicitly
have to use a context manager (which is preferred, anyway), otherwise,
they have to call `file.close()` themselves.