Fix failing client initialization test
It turns out that passing kwargs at all (even if they are `None`) will
prompt lakeFS to use them, and not attempt filling the `None`s with
values from other initialization methods.

To fix, fall back to the old way of checking all client configuration
arguments to be `None`, and if so, call the `lakefs.Client()` constructor
without any arguments.

Adds back the previous SDK client configuration arguments to the
constructor.