Remove `lakefs_spec.config` module
This functionality has been integrated in the `lakefs` SDK wrapper
library. It currently only allows configuration files sourced from
`$HOME/.lakectl.yaml`, but we can just add a configfile argument again
if it is changed upstream in the future.

Drops `pyyaml` from the project dependencies, since `lakefs` depends on
`pyyaml` for the same reason we did, which is reading in lakectl files.