Use `lakefs.Reference`s everywhere for object infos
The methods are available on the base classes, which means they should
be able to be used instead of distinct branch names.

However, test coverage might need to be improved, as I do not recall
implementing a test for a commit input on any `stat`-like API.Use `lakefs.Reference`s everywhere for object infos

The methods are available on the base classes, which means they should
be able to be used instead of distinct branch names.

However, test coverage might need to be improved, as I do not recall
implementing a test for a commit input on any `stat`-like API.