Implement file transfers via lakefs.object.Object(Reader|Writer)
Since the vital methods needed for transactions were only included in
v0.2.0, this change incurs a mandatory upgrade to lakefs v0.2.0.

The implicit branch creation is included in the `mode != "rb"` branch
in `fs._open()`.

The migration required the overload of `tail()`, since the `Reader` and
`Writer` do not have a size attribute.