Remove Placeholder from interface typing, use new ReferenceType hint
As a consequence of my complaints upstream, a new `ReferenceType` was
added, which is the union of `str` and everything that marks a reference
in lakefs Python code (branches, tags, commits). We now use this
everywhere, as it saves space in the interfaces.

Also, since the placeholder is ducktyped to be a ReferenceType always,
and passes type checks at runtime (isinstance only), the typing could
be relaxed to work with the native reference type only.