Rework transaction APIs
Since the repository and ephemeral branch are now saved on the
transaction, they are no longer needed on the APIs. Commits through
`LakeFSTransaction.commit()` are automatically created on the ephemeral
branch.

The biggest change is for rev_parse, which does not need the complicated
commit fetching anymore, and can just take the ref expression directly.