Migrate tests to new transaction API
Requires repository, base branch everywhere.

Right now, we solve all file puts by interpolating the string in-place
with `tx.branch` instead of `temp_branch`. This is relatively ugly, and
could be better with transaction APIs that change state on the branch
(puts, rms).
However, those need a check that the input path can be
relative, since we just by default assume the user wants to put on the
ephemeral branch.