Create new branch before transaction
Previously, the new branch was a result of implicit branch creation in
`fs.put_file()` _within_ the context.

Now, since we do not allow the base branch to be created in situ, which
would be very confusing, we simply create it explicitly before the
transaction start.

We do this for both NEW_BRANCH (not sure what it is used for in the end)
and TRAINING_BRANCH.