Introduce new transaction model based on ephemeral branches (#258)
This is a breaking change, as it requires users to give arguments to the return object of `LakeFSFileSystem.transaction`.

Changes the `LakeFSTransaction` class to operate on a repository+branch-scoped ephemeral branch.

File uploads, removals, and commits happen on that ephemeral branch. After transaction exit, the created branch is optionally merged back into the base branch.

The base branch (from which we branch off in the transaction) must exist before start of the transaction.

Some transaction APIs (merge, revert, tag) can also work on branches other than the transaction's ephemeral branch - this behavior might not be well-defined, and is up for debate on removal.

Similarly, on puts and deletes, one currently has to specify the transaction branch directly via the `LakeFSTransaction.branch` property. This can be made nicer by allowing resource paths relative to the transaction's repo and branch in special functions like e.g. `LakeFSTransaction.put()` (deferring to `fs.put()`), `LakeFSTransaction.rm()` (deferring to `fs.rm()`), and so on.

This pull request turns the LakeFSTransaction into a callable, which needs to be called with a repository name at minimum to ensure transaction execution.