Always show the project name in the header
When scrolling down on a documentation page, the page header will change
from showing the project title (i.e. lakeFS-spec) to showing the current
page title.

From that point on there is no more visual indication that you are
viewing the lakeFS-spec documentation.

This change will prepend the project title to the page title.