fix: Attempt to fix Google canonical URL detection
It seems that a meta redirect with a timeout greater than 0
is treated as a temporary redirect by the Google crawler.

This seems to cause a discrepancy between our reported
canonical site URLs and the ones inferred by Google.

Setting the timeout to 0 may force Google to update its view
of the canonical URLs.