Batch objects in `Branch.delete_objects` calls (#285)
lakeFS supports a maximum of 1000 objects per `delete_objects()`
invocation, since that calls the object deletion API under the hood.
This limitation is not without precedent, and exists also e.g. on
AWS S3.

Hence, we batch delete objects on the client side, dispatching multiple
API calls in the cases where the deleted objects list contains more
than a thousand objects.

The threshold number of objects is currently hardcoded to 1000, since
that is the limitation of the lakeFS server.