Improve error message in `util.parse()` by incrementally using regexes
This loses a bit of performance, but allows us to improve the error message
to the user by attaching information about which part of the URI was invalid.

Since we consider the regexes an implementation detail, we do not return
them in the error message.

Changes the failure test to include the expected invalid parts to harden the test case.