Remove TestPyPI upload due to attestation failure
A generated attestation in the upload job can only be used for a single upload,
and is not cleared between uploads - hence, the PyPI release workflow reuses the
attestation of TestPyPI, and gets rejected.

For now, remove the TestPyPI upload, so that we don't run into the stale attestation
problem.