refactor: Move request config to `lakefs_spec.types`, pass request_config explicitly
The test was hacky in that the `request_config` was passed by setting it as an attribute
later, which meant you had to preface keys with an underscore yourself.
Now we just pass it into the constructor with non-private keys, in the intended way we
describe in the docs.

NB: To satisfy mypy, I had to explicitly put a `RequestConfig` annotation.