feat(fs): Pass kwargs to SDK object wrapper calls (#330)
* feat: Add `request_config` slot to `LakeFSFileSystem` constructor

This field takes a `RequestConfig` object, which is a typed dict containing the
customizable keyword arguments of the lakeFS API with their best-effort types.
It is marked as non-total, so that a partial configuration (e.g. only the timeout)
is supported.

* docs: Add note on configuring API kwargs with usage example

Contains the specification verbatim, to document all possible values for API request
configuration.

* refactor: Move request config to `lakefs_spec.types`, pass request_config explicitly

The test was hacky in that the `request_config` was passed by setting it as an attribute
later, which meant you had to preface keys with an underscore yourself.
Now we just pass it into the constructor with non-private keys, in the intended way we
describe in the docs.

NB: To satisfy mypy, I had to explicitly put a `RequestConfig` annotation.

* docs: Cut request config definition out of usage example, prefer import

This way, it is less confusing to users, and shows the simplest way of using the config
type, which is by just importing it from the `lakefs_spec.types` submodule.