feat(transaction): Support finer merge customization through kwargs
Instead of `squash`, define a `MergeKwargs` type, holding all possible merge
keyword argument types supported by the lakeFS API client.

Merge kwargs are given on transaction entry, and are applied in the merge on close
if and only if `automerge` is set to true, and no errors occurred in the transaction scope.