chore(packaging): Explicitly depend on `typing_extensions` for Python 3.10
Previously, we were using `typing_extensions` members freely even though we do not
explicitly depend on them, which only worked because of a chain of implicit dependencies
(lakefs -> lakefs-sdk -> pydantic -> typing-extensions).

Now, we change it to the more explicit "depend on what you use" strategy, and add
`typing_extensions` to the dependencies. Since we only use extensions added in Python 3.11,
we restrict the dependency to Python 3.10 only.