fix(fs): Use `autocommit` value from signature instead of computing it (#334)
Unlike fsspec's base class, which infers the value of `autocommit` from the kwargs
and, if necessary, from the current transaction state, we keep `autocommit` as an
explicit named keyword argument in our signature of `LakeFSFileSystem.open()`.

Previously, this value was unused, and the keyword arguments would never
contain autocommit due to name collision, so we implicitly autocommitted files if and
only if we were not in a transaction. This change makes the file committing strategy
configurable, and defaults to `False`.