feat(transaction): Support finer merge customization through kwargs (#335)
* feat(transaction): Support finer merge customization through kwargs

Instead of `squash`, define a `MergeKwargs` type, holding all possible merge
keyword argument types supported by the lakeFS API client.

Merge kwargs are given on transaction entry, and are applied in the merge on close
if and only if `automerge` is set to true, and no errors occurred in the transaction scope.

Previously, we were using `typing_extensions` members freely even though we do not
explicitly depend on them, which only worked because of a chain of implicit dependencies
(lakefs -> lakefs-sdk -> pydantic -> typing-extensions).

Now, we change it to the more explicit "depend on what you use" strategy, and add
`typing_extensions` to the dependencies.

* docs(transactions): Add note on merge kwargs customization

Follows the style of the recently added `request_config` doc.
Contains a definition of the `MergeKwargs` struct, as well as a code example.

* deps: Depend unconditionally on typing-extensions

Remove import guards, since typing-extensions does the same with typing under
the hood.